# lib/api-zod/src/generated/types/playerSosRow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StatChasers Fantasy Football SOS API
 * OpenAPI spec version: 0.1.0
 */
import type { WeekCell } from './weekCell';

export interface PlayerSosRow {
  playerId: string;
  playerName: string;
  team: string;
  position: string;
  sosRank: number;
  playoffSosRank: number;
  rosSosRank: number;
  weeks: WeekCell[];
  playoff2?: WeekCell;
  playoff3?: WeekCell;
}