# lib/api-zod/src/generated/types/scheduleResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StatChasers Fantasy Football SOS API
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleGame } from './scheduleGame';

export interface ScheduleResponse {
  season: number;
  games: ScheduleGame[];
}