# lib/api-zod/src/generated/types/teamSosRow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StatChasers Fantasy Football SOS API
 * OpenAPI spec version: 0.1.0
 */
import type { WeekCell } from './weekCell';

export interface TeamSosRow {
  team: string;
  teamFullName: string;
  division: string;
  overallRank: number;
  rosRank: number;
  playoffRank: number;
  weeks: WeekCell[];
  playoff2?: WeekCell;
  playoff3?: WeekCell;
  rosSummary?: number | null;
}